Moved credentials-prompt to proper place
