+++ lib/cli/publish.js
@@ -1,29 +1,9 @@
 #! /usr/bin/env node
 
 var program = require('commander');
-var cliPrompt = require('prompt');
 
 var dasherize = require('../utils/string').dasherize;
 var ui = require('../ui');
-
-cliPrompt.message = '';
-cliPrompt.delimiter = '';
-
-var promptSchema = {
-  properties: {
-    username: {
-      description: 'GitHub username: ',
-      message: 'A username is required.',
-      required: true
-    },
-    password: {
-      description: 'GitHub password: ',
-      message: 'A password is required.',
-      required: true,
-      hidden: true
-    }
-  }
-};
 
 program
   .arguments('<addonName>')
@@ -48,20 +28,7 @@ function runPublishTask() {
 }
 
 if (program.addonName) {
-  if(program.username && program.password) {
-    runPublishTask();
-  } else {
-    cliPrompt.start();
-    cliPrompt.get(promptSchema, function(error, result) {
-      if (error) {
-        return ui.error(error);
-      } else {
-        program.username = result.username;
-        program.password = result.password;
-        runPublishTask();
-      }
-    });
-  }
+  runPublishTask();
 } else {
   program.outputHelp();
 }
