+++ lib/tasks/publish.js
@@ -8,12 +8,10 @@ var GitHubAPI = require('github');
 var Promise = require('../ext/promise');
 var ui = require('../ui');
 var gitHubCredentials = require('../utils/github-credentials');
+var promptForCredentials = require('../utils/credentials-prompt');
 
-function createGitSignature(options) {
-  var millisecondsFromStandardEpoch = Date.now();
-  var secondsFromStandardEpoch = Math.floor(millisecondsFromStandardEpoch / 1000);
-
-  return git.Signature.create('ember-micro', options.username, secondsFromStandardEpoch, 60);
+function createGitSignature(repo) {
+  return git.Signature.default(repo);
 }
 
 function createLocalRepo(options) {
@@ -43,8 +41,9 @@ function addExistingFilesToLocalRepo(localRepo) {
 }
 
 function createInitialCommitOnLocalRepo(options, localRepo, oid) {
-  var signature = createGitSignature(options);
-
+  console.log('creating signature', options);
+  var signature = createGitSignature(localRepo);
+  console.log('signature created', signature);
   return localRepo.createCommit("HEAD", signature, signature, "Initial commit", oid, []);
 }
 
@@ -73,7 +72,7 @@ function pushToRepo(options, localRepo, remote) {
 
   remote.setCallbacks({
     credentials: function() {
-      return git.Cred.userpassPlaintextNew(options.username, options.password);
+      return git.Cred.usernameNew('ember-micro');
     }
   });
 
@@ -90,17 +89,24 @@ function initializeGitHub(options) {
     version: '3.0.0',
   });
 
-  return gitHubCredentials.getToken().then(function(token) {
-    console.log('token exists:', token);
-    authenticateWithToken(github, token);
-    return github;
-  }).catch(function() {
-    console.log('token missing, creating');
+  if (options.username && options.password) {
     authenticateWithUserProvidedCredentials(github,options);
     return createToken(github).then(gitHubCredentials.setToken).then(function() {
       return github;
     });
-  });
+  } else {
+    return gitHubCredentials.getToken().then(function(token) {
+      authenticateWithToken(github, token);
+      return github;
+    }).catch(function() {
+      return promptForCredentials().then(function(credentials) {
+        authenticateWithUserProvidedCredentials(github, credentials);
+        return createToken(github).then(gitHubCredentials.setToken).then(function() {
+          return github;
+        });
+      });
+    });
+  }
 }
 
 function authenticateWithToken(github, token) {
@@ -122,7 +128,7 @@ function createToken(github) {
   return new Promise(function(resolve, reject) {
     github.authorization.create({
       scopes: ['user', 'public_repo', 'repo', 'repo:status', 'gist'],
-      note: 'ember-micro-addon authorization',
+      note: 'ember-micro-addon@' + Date.now(),
       note_url: 'https://github.com/coderly/ember-micro-addon',
       headers: {
         'X-GitHub-OTP': 'two-factor-code'
